Migration workaround for #13
This is part of a bloody hack for
https://github.com/JamborJan/paperwork/issues/13

We show an export option in when we detect that an update is due from the Paperwork Sandstorm package V3 to V5 (V4 never existed). The user then has to export the data and import it again in an empty instance.

This hack also requires this change in the not git monitored file

/opt/app/frontend/vendor/laravel/framework/src/Illuminate/Database/Migrations/Migrator.php

https://gist.github.com/JamborJan/767f1dc8bede9f97dd53#file-migrator-php-L108