Adjusted Base Box
Solved this one:

*** INCOMPATIBLE OPTIONS DETECTED ***

You need to adjust .sandstorm/Vagrantfile due to recent incompatible changes
made by a project we depend on. Apologies for this mess.

Here's how to fix it:

Open .sandstorm/Vagrantfile and find this line:

  config.vm.box = "debian/jessie64"

Change it to:

  config.vm.box = "sandstorm/debian-jessie64"

For details (including how to bypass this warning) read:
https://github.com/sandstorm-io/vagrant-spk/releases/tag/v0.137

Once you have done that, please re-run this command ("vagrant-spk up").